feat: semantic element for author bios
