+++ styles/global-blocks.css
@@ -22,15 +22,16 @@
   Where necessary, `:root` selectors are located near their related blocks.
 */
 
+.author-aside {
+  margin-block-start: 2rem;
+  border-top: 1px solid var(--spectrum-gray-400);
+}
+
 .author-bio {
   margin-block-start: 2rem;
   margin-block-end: 0;
-  padding-block-start: 2rem;
-  border-top: 1px solid var(--spectrum-gray-400);
 
   + .author-bio {
-    border-top: none;
-    padding-block-start: 0;
     margin-block-start: 1rem;
   }
 
@@ -52,6 +53,8 @@
       gap: 0.25rem;
 
       > .author-meta__name {
+        margin-block: 0;
+
         @media (min-width: 48rem) {
           font-size: var(--spectrum-font-size-500);
         }
